feat: Add base CRUD operations for Position entity, create PositionController, PositionValidator, change PositionRepository, PositionService, PositionDTO, DepartmentControler. Create template for pagination, create templates for positions, change employee.html, department.html, companyBranch.html, header.html

+++ src/main/java/com/manageemployee/employeemanagement/service/PositionService.java
@@ -3,9 +3,13 @@ package com.manageemployee.employeemanagement.service;
 import com.manageemployee.employeemanagement.model.Position;
 import com.manageemployee.employeemanagement.repository.PositionRepository;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
+import java.util.Optional;
 
 @Service
 public class PositionService {
@@ -16,19 +20,38 @@ public class PositionService {
         this.positionRepository = positionRepository;
     }
 
-    public Position createOrUpdatePosition(Position position) {
+    @Transactional
+    public Position createPosition(Position position) {
+        return positionRepository.saveAndFlush(position);
+    }
+
+    @Transactional
+    public Position updatePosition(Position position) {
         return positionRepository.saveAndFlush(position);
     }
 
     public Position getPositionById(Long id) {
-        return positionRepository.findById(id).orElse(null);
+        return positionRepository.findById(id).orElseThrow(() ->
+                new IllegalArgumentException("Выбрана не существующая должность!"));
     }
 
-    public List<Position> getAllPositions() {
-        return positionRepository.findAll();
+    public Page<Position> getPositionsByDepartment(Pageable pageable, Long departmentId) {
+        return positionRepository.findAllByDepartment_Id(pageable, departmentId);
+    }
+
+    public boolean existsByNameAndDepartmentId(String positionName, Long depId) {
+        return positionRepository.existsByPositionNameIgnoreCaseAndDepartment_Id(positionName, depId);
+    }
+
+    public Optional<Position> findByNameAndDepartmentId(String positionName, Long depId) {
+        return positionRepository.findByPositionNameIgnoreCaseAndDepartment_Id(positionName, depId);
     }
 
     public void deletePositionById(Long id) {
+        Position position = getPositionById(id);
+        if (position.getAmountOfEmployees() != 0)
+            throw new UnsupportedOperationException("Нельзя удалить должность на которой числятся сотрудники!");
+
         positionRepository.deleteById(id);
     }
 }
